requirements.txt
locustio==0.8.1
cryptography ~= 2.3
pyjwt==1.6.0
paho-mqtt==1.3.1

